refactor controllers
