+++ endpoint/src/main/java/ru/studentsplatform/backend/endpoint/rest/crud/DirectionController.java
@@ -1,4 +1,4 @@
-package ru.studentsplatform.backend.endpoint.rest;
+package ru.studentsplatform.backend.endpoint.rest.crud;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
