+++ endpoint/src/main/java/ru/studentsplatform/backend/endpoint/rest/crud/impl/UserControllerImpl.java
@@ -1,11 +1,11 @@
-package ru.studentsplatform.backend.endpoint.rest.crud;
+package ru.studentsplatform.backend.endpoint.rest.crud.impl;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import ru.studentsplatform.backend.domain.dto.user.UserDTO;
 import ru.studentsplatform.backend.endpoint.mapper.UserMapper;
-import ru.studentsplatform.backend.endpoint.rest.UserConroller;
+import ru.studentsplatform.backend.endpoint.rest.crud.UserConroller;
 import ru.studentsplatform.backend.service.crud.impl.UserServiceImpl;
 import ru.studentsplatform.backend.system.annotation.Profiled;
